+++ Client/src/init.js
@@ -594,5 +594,5 @@ export default appMgr => {
     }
   });
 
-  appMgr.uiStateMgr.setActivePageId(5, 0);
+  appMgr.uiStateMgr.setActivePageId(1, 0);
 };